Move data classes to models package
