+++ app/src/main/java/com/example/traders/network/models/CryptoTicker.kt
@@ -1,4 +1,4 @@
-package com.example.traders.profile.cryptoData
+package com.example.traders.network.models
 
 data class CryptoTicker(
     val symbol: String,
