+++ app/src/main/java/com/example/traders/profile/adapters/PortfolioListAdapter.kt
@@ -9,7 +9,7 @@ import com.bumptech.glide.Glide
 import com.example.traders.R
 import com.example.traders.databinding.ListItemCryptoBinding
 import com.example.traders.databinding.ListPortfolioHeaderBinding
-import com.example.traders.profile.cryptoData.CryptoInUsd
+import com.example.traders.profile.portfolio.CryptoInUsd
 import com.example.traders.presentation.watchlist.adapters.SingleCryptoListener
 import com.example.traders.database.FixedCryptoList
