+++ app/src/main/java/com/example/traders/profile/portfolio/PortfolioState.kt
@@ -1,8 +1,7 @@
 package com.example.traders.profile.portfolio
 
 import com.example.traders.database.Crypto
-import com.example.traders.profile.cryptoData.CryptoInUsd
-import com.example.traders.profile.cryptoData.CryptoTicker
+import com.example.traders.network.models.CryptoTicker
 import com.github.mikephil.charting.data.PieEntry
 import java.math.BigDecimal
