+++ app/src/test/java/com/example/traders/profile/portfolio/PortfolioViewModelTest.kt
@@ -6,8 +6,7 @@ import androidx.lifecycle.MutableLiveData
 import app.cash.turbine.test
 import com.example.traders.TestDispatcherRule
 import com.example.traders.database.Crypto
-import com.example.traders.profile.cryptoData.CryptoInUsd
-import com.example.traders.profile.cryptoData.CryptoTicker
+import com.example.traders.network.models.CryptoTicker
 import com.example.traders.network.repository.CryptoRepository
 import com.example.traders.utils.roundNum
 import com.github.mikephil.charting.data.PieData
